requirements.txt
aws-cdk-lib>=2.38.0
constructs>=10.0.0
boto3==1.40.14
botocore==1.40.14
PyYAML
